role assignment support added
