+++ lib/azure/azHttpClient.js
@@ -34,6 +34,7 @@ const azAuthentication_1 = require("./azAuthentication");
 const httpClient_1 = require("../utils/httpClient");
 const policyHelper_1 = require("./policyHelper");
 const utilities_1 = require("../utils/utilities");
+const roleAssignmentHelper_1 = require("./roleAssignmentHelper");
 const SYNC_BATCH_CALL_SIZE = 20;
 const DEFINITION_SCOPE_SEPARATOR = "/providers/Microsoft.Authorization/policyDefinitions";
 class AzHttpClient {
@@ -128,9 +129,12 @@ class AzHttpClient {
             return this.upsertPolicies(policyRequests);
         });
     }
-    upsertPolicyAssignments(policyRequests) {
+    upsertPolicyAssignments(policyRequests, roleAssignmentResults) {
         return __awaiter(this, void 0, void 0, function* () {
-            return this.upsertPolicies(policyRequests);
+            const assignmentResponses = yield this.upsertPolicies(policyRequests);
+            // Now we need to add roles to managed identity for policy remediation.
+            yield roleAssignmentHelper_1.assignRoles(policyRequests, assignmentResponses, roleAssignmentResults);
+            return assignmentResponses;
         });
     }
     deletePolicies(policyIds) {
