+++ lib/azure/forceUpdateHelper.js
@@ -65,9 +65,10 @@ function handleForceUpdate(definitionRequests, policyResponses, assignmentReques
                 // Get all assignments in one array
                 let assignmentsInService = [].concat(...allDefinitionAssignments);
                 let repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses;
+                let roleAssignmentResults = [];
                 // Start force update.
                 try {
-                    [repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses] = yield startForceUpdate(badRequests, definitionsInService, assignmentsInService, azHttpClient);
+                    [repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses, roleAssignmentResults] = yield startForceUpdate(badRequests, definitionsInService, assignmentsInService, azHttpClient);
                 }
                 catch (error) {
                     utilities_1.prettyLog(`ForceUpdate Failed. Error : ${error}`);
@@ -78,6 +79,7 @@ function handleForceUpdate(definitionRequests, policyResponses, assignmentReques
                 removeAssignmentRequests(assignmentRequests, repoAssignmentResponses);
                 // Populate results
                 populateResults(badRequests, repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses, policyResults);
+                policyResults.push(...roleAssignmentResults);
                 utilities_1.prettyLog("ForceUpdate : End");
             }
             else {
@@ -93,9 +95,10 @@ exports.handleForceUpdate = handleForceUpdate;
 function startForceUpdate(badRequests, definitionsInService, assignmentsInService, azHttpClient) {
     return __awaiter(this, void 0, void 0, function* () {
         let duplicateDefinitions, duplicateAssignments;
+        let duplicateroleAssignmentResults = [];
         // Duplicate definitions and assignments in Azure before deletion
         try {
-            [duplicateDefinitions, duplicateAssignments] = yield createDuplicatePolicies(definitionsInService, assignmentsInService, azHttpClient);
+            [duplicateDefinitions, duplicateAssignments] = yield createDuplicatePolicies(definitionsInService, assignmentsInService, duplicateroleAssignmentResults, azHttpClient);
         }
         catch (error) {
             console.log(`Error occurred while creating duplicate policies. Abandoning force update. Error : ${error}`);
@@ -118,8 +121,9 @@ function startForceUpdate(badRequests, definitionsInService, assignmentsInServic
         }
         // Create fresh policies from repo
         let repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses;
+        let roleAssignmentResults = [];
         try {
-            [repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses] = yield createPoliciesFromCode(badRequests, azHttpClient);
+            [repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses] = yield createPoliciesFromCode(badRequests, roleAssignmentResults, azHttpClient);
         }
         catch (error) {
             console.log(`Error occurred while creating policies from code. Abandoning force update. Error : ${error}`);
@@ -141,14 +145,15 @@ function startForceUpdate(badRequests, definitionsInService, assignmentsInServic
             console.log(`Could not delete duplicate policies.`);
             // TODO PM: what to do now? policies are updated but duplicates could not be deleted.
         }
-        return [repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses];
+        return [repoDefinitionResponses, repoAssignmentRequests, repoAssignmentResponses, roleAssignmentResults];
     });
 }
 function revertOldPoliciesAndDeleteDuplicates(definitions, assignments, duplicateDefinitions, duplicateAssignments, azHttpClient) {
     return __awaiter(this, void 0, void 0, function* () {
         console.log("Reverting to old policies and deleting duplicates.");
+        let roleAssignmentResults = [];
         try {
-            yield upsertOldPolicies(definitions, assignments, azHttpClient);
+            yield upsertOldPolicies(definitions, assignments, roleAssignmentResults, azHttpClient);
         }
         catch (error) {
             console.log(`Could not revert to old policies.`);
@@ -211,12 +216,12 @@ function getAllDefinitionsAssignment(policyDefinitionIds, azHttpClient) {
         return responses.map(response => response.content.value);
     });
 }
-function createDuplicatePolicies(policyDefinitions, policyAssignments, azHttpClient) {
+function createDuplicatePolicies(policyDefinitions, policyAssignments, roleAssignmentResults, azHttpClient) {
     return __awaiter(this, void 0, void 0, function* () {
         core.debug('Force Update : Creating Duplicate Policies');
         const duplicateDefinitionRequests = createDuplicateRequests(policyDefinitions);
         const duplicateAssignmentRequests = createDuplicateRequests(policyAssignments);
-        const [duplicateDefinitionsResponses, duplicateAssignmentsResponses] = yield createPolicies(duplicateDefinitionRequests, duplicateAssignmentRequests, azHttpClient, true);
+        const [duplicateDefinitionsResponses, duplicateAssignmentsResponses] = yield createPolicies(duplicateDefinitionRequests, duplicateAssignmentRequests, roleAssignmentResults, azHttpClient, true);
         return [getPoliciesFromResponse(duplicateDefinitionsResponses), getPoliciesFromResponse(duplicateAssignmentsResponses)];
     });
 }
@@ -277,22 +282,22 @@ function deletePolicies(policies, azHttpClient) {
 /**
  * Reverts policies in service which were deleted/modified.
  */
-function upsertOldPolicies(policyDefinitions, policyAssignments, azHttpClient) {
+function upsertOldPolicies(policyDefinitions, policyAssignments, roleAssignmentResults, azHttpClient) {
     return __awaiter(this, void 0, void 0, function* () {
         const definitionRequests = getPolicyRequests(policyDefinitions);
         const assignmentRequests = getPolicyRequests(policyAssignments);
-        yield createPolicies(definitionRequests, assignmentRequests, azHttpClient);
+        yield createPolicies(definitionRequests, assignmentRequests, roleAssignmentResults, azHttpClient);
     });
 }
 /**
  * Creates definition, corresponging assignments which needed force update.
  * In case of failure, created policies are deleted.
  */
-function createPoliciesFromCode(definitionRequests, azHttpClient) {
+function createPoliciesFromCode(definitionRequests, roleAssignmentResults, azHttpClient) {
     return __awaiter(this, void 0, void 0, function* () {
         utilities_1.prettyDebugLog(`Force update : Creating policies in code`);
         const assignmentRequests = getAssignmentRequests(definitionRequests);
-        const [definitionResponses, assignmentResponses] = yield createPolicies(definitionRequests, assignmentRequests, azHttpClient, true);
+        const [definitionResponses, assignmentResponses] = yield createPolicies(definitionRequests, assignmentRequests, roleAssignmentResults, azHttpClient, true);
         return [definitionResponses, assignmentRequests, assignmentResponses];
     });
 }
@@ -300,11 +305,11 @@ function createPoliciesFromCode(definitionRequests, azHttpClient) {
  * Creates policy definitions and assignments. Throws in case any policy creation fails.
  * In case there is a failure while creation and 'deleteInFailure' is true then all policies are deleted.
  */
-function createPolicies(definitionRequests, assignmentRequests, azHttpClient, deleteInFailure = false) {
+function createPolicies(definitionRequests, assignmentRequests, roleAssignmentResults, azHttpClient, deleteInFailure = false) {
     return __awaiter(this, void 0, void 0, function* () {
         const definitionResponses = yield azHttpClient.upsertPolicyDefinitions(definitionRequests);
         yield validateUpsertResponse(definitionResponses, azHttpClient, deleteInFailure);
-        const assignmentResponses = yield azHttpClient.upsertPolicyAssignments(assignmentRequests);
+        const assignmentResponses = yield azHttpClient.upsertPolicyAssignments(assignmentRequests, roleAssignmentResults);
         try {
             yield validateUpsertResponse(assignmentResponses, azHttpClient, deleteInFailure);
         }
