+++ lib/azure/policyHelper.js
@@ -28,10 +28,9 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.createPoliciesUsingIds = exports.getPolicyResults = exports.getPolicyAssignment = exports.getPolicyAssignments = exports.createUpdatePolicies = exports.getAllPolicyRequests = exports.FRIENDLY_ASSIGNMENT_TYPE = exports.FRIENDLY_INITIATIVE_TYPE = exports.FRIENDLY_DEFINITION_TYPE = exports.POLICY_INITIATIVE_FILE_NAME = exports.POLICY_FILE_NAME = exports.POLICY_RESULT_SUCCEEDED = exports.POLICY_RESULT_FAILED = exports.POLICY_OPERATION_NONE = exports.POLICY_OPERATION_UPDATE = exports.POLICY_OPERATION_CREATE = exports.ROLE_ASSIGNMNET_TYPE = exports.ASSIGNMENT_TYPE = exports.INITIATIVE_TYPE = exports.DEFINITION_TYPE = void 0;
+exports.createPoliciesUsingIds = exports.isCreateOperation = exports.getPolicyResults = exports.getPolicyAssignment = exports.getPolicyAssignments = exports.createUpdatePolicies = exports.getAllPolicyRequests = exports.FRIENDLY_ASSIGNMENT_TYPE = exports.FRIENDLY_INITIATIVE_TYPE = exports.FRIENDLY_DEFINITION_TYPE = exports.POLICY_INITIATIVE_FILE_NAME = exports.POLICY_FILE_NAME = exports.POLICY_RESULT_SUCCEEDED = exports.POLICY_RESULT_FAILED = exports.POLICY_OPERATION_NONE = exports.POLICY_OPERATION_UPDATE = exports.POLICY_OPERATION_CREATE = exports.ROLE_ASSIGNMNET_TYPE = exports.ASSIGNMENT_TYPE = exports.INITIATIVE_TYPE = exports.DEFINITION_TYPE = void 0;
 const path = __importStar(require("path"));
 const core = __importStar(require("@actions/core"));
-const httpClient_1 = require("../utils/httpClient");
 const azHttpClient_1 = require("./azHttpClient");
 const forceUpdateHelper_1 = require("./forceUpdateHelper");
 const fileHelper_1 = require("../utils/fileHelper");
@@ -39,7 +38,6 @@ const hashUtils_1 = require("../utils/hashUtils");
 const utilities_1 = require("../utils/utilities");
 const pathHelper_1 = require("../inputProcessing/pathHelper");
 const Inputs = __importStar(require("../inputProcessing/inputs"));
-const uuid_1 = require("uuid");
 exports.DEFINITION_TYPE = "Microsoft.Authorization/policyDefinitions";
 exports.INITIATIVE_TYPE = "Microsoft.Authorization/policySetDefinitions";
 exports.ASSIGNMENT_TYPE = "Microsoft.Authorization/policyAssignments";
@@ -111,10 +109,8 @@ function createUpdatePolicies(policyRequests) {
         policyResults.push(...getPolicyResults(definitionRequests, definitionResponses, exports.FRIENDLY_DEFINITION_TYPE));
         const initiativeResponses = yield azHttpClient.upsertPolicyInitiatives(initiativeRequests);
         policyResults.push(...getPolicyResults(initiativeRequests, initiativeResponses, exports.FRIENDLY_INITIATIVE_TYPE));
-        const assignmentResponses = yield azHttpClient.upsertPolicyAssignments(assignmentRequests);
+        const assignmentResponses = yield azHttpClient.upsertPolicyAssignments(assignmentRequests, policyResults);
         policyResults.push(...getPolicyResults(assignmentRequests, assignmentResponses, exports.FRIENDLY_ASSIGNMENT_TYPE));
-        // Now we need to add roles to managed identity for policy remediation.
-        yield assignRoles(assignmentRequests, assignmentResponses, policyResults);
         return Promise.resolve(policyResults);
     });
 }
@@ -139,132 +135,6 @@ function dividePolicyRequests(policyRequests) {
         }
     });
     return [definitionRequests, initiativeRequests, assignmentRequests];
-}
-function assignRoles(assignmentRequests, assignmentResponses, roleAssignmentResults) {
-    return __awaiter(this, void 0, void 0, function* () {
-        let filteredAssignments = filterIdentityAssignments(assignmentRequests, assignmentResponses);
-        let allRoleDefinitions = yield paresRoleDefinitions(filteredAssignments, roleAssignmentResults);
-        let roleRequests = getRoleRequests(filteredAssignments, allRoleDefinitions);
-        yield createRoles(roleRequests, roleAssignmentResults);
-    });
-}
-function filterIdentityAssignments(assignmentRequests, assignmentResponses) {
-    let filteredAssignments = [];
-    assignmentRequests.forEach((assignmentRequest, index) => {
-        let assignmentResponse = assignmentResponses[index];
-        // We will assign roles only when assignmnet was created and has identity field has principalId in it.
-        if (isCreateOperation(assignmentRequest) && assignmentResponse.identity && assignmentResponse.identity.principalId) {
-            // We will add path in assignment as it is required later.
-            assignmentResponse.path = assignmentRequest.path;
-            filteredAssignments.push(assignmentResponse);
-        }
-    });
-    return filteredAssignments;
-}
-function paresRoleDefinitions(policyAssignments, roleAssignmentResults) {
-    return __awaiter(this, void 0, void 0, function* () {
-        let roleDefinitions = {};
-        const policyDefinitionIds = policyAssignments.map(assignment => assignment.properties.policyDefinitionId);
-        try {
-            const azHttpClient = new azHttpClient_1.AzHttpClient();
-            yield azHttpClient.initialize();
-            let policyDefinitions = yield azHttpClient.getPolicyDefintions(policyDefinitionIds);
-            policyDefinitions.forEach((definition, index) => {
-                if (definition.error) {
-                    let policyAssignment = policyAssignments[index];
-                    let message = definition.error.message ? definition.error.message : "Could not get policy definition from Azure";
-                    roleAssignmentResults.push(getRoleAssignmentResult(policyAssignment.path, policyAssignment.id, policyAssignment.properties.policyDefinitionId, exports.POLICY_RESULT_FAILED, message));
-                }
-                else {
-                    let roleDefinitionIds = getRoleDefinitionIds(definition);
-                    if (roleDefinitionIds && roleDefinitionIds.length > 0) {
-                        // We need last part of role definition id 
-                        roleDefinitions[definition.id] = roleDefinitionIds.map(roleDefinitionId => roleDefinitionId.split("/").pop());
-                    }
-                    else {
-                        utilities_1.prettyLog(`Could not find role definition ids for adding role assignments to the managed identity. Definition Id : ${definition.id}`);
-                    }
-                }
-            });
-        }
-        catch (error) {
-            utilities_1.prettyDebugLog(`An error occurred while getting role requests for missing policy definitions. Error : ${error}`);
-            throw new Error(`An error occurred while getting role requests for missing policy definitions. Error: ${error}`);
-        }
-        return roleDefinitions;
-    });
-}
-function getRoleAssignmentResult(path, assignmentId, definitionId, status, message) {
-    return {
-        path: path,
-        type: exports.ROLE_ASSIGNMNET_TYPE,
-        operation: exports.POLICY_OPERATION_CREATE,
-        displayName: `Role Assignment for policy policy assignment id : ${assignmentId}`,
-        status: status,
-        message: message,
-        policyDefinitionId: definitionId
-    };
-}
-function getRoleRequests(policyAssignments, allRoleDefinitions) {
-    let roleRequests = [];
-    policyAssignments.forEach(policyAssignment => {
-        let roleDefinitions = allRoleDefinitions[policyAssignment.properties.policyDefinitionId];
-        if (roleDefinitions) {
-            roleDefinitions.forEach(roleId => {
-                roleRequests.push({
-                    scope: policyAssignment.properties.scope,
-                    roleAssignmentId: uuid_1.v4(),
-                    roleDefinitionId: roleId,
-                    principalId: policyAssignment.identity.principalId,
-                    policyAssignmentId: policyAssignment.id,
-                    policyDefinitionId: policyAssignment.properties.policyDefinitionId,
-                    path: policyAssignment.path
-                });
-            });
-        }
-    });
-    return roleRequests;
-}
-function createRoles(roleRequests, roleAssignmentResults) {
-    return __awaiter(this, void 0, void 0, function* () {
-        if (roleRequests.length == 0) {
-            utilities_1.prettyDebugLog(`No role assignments needs to be created`);
-            return;
-        }
-        try {
-            const azHttpClient = new azHttpClient_1.AzHttpClient();
-            yield azHttpClient.initialize();
-            let responses = yield azHttpClient.addRoleAssinments(roleRequests);
-            responses.forEach((response, index) => {
-                let roleRequest = roleRequests[index];
-                let message = `Role Assignment created with id : ${response.content.id}`;
-                let status = exports.POLICY_RESULT_SUCCEEDED;
-                if (response.httpStatusCode == httpClient_1.StatusCodes.CREATED) {
-                    utilities_1.prettyDebugLog(`Role assignment created with id ${response.content.id} for assignmentId : ${roleRequest.policyAssignmentId}`);
-                }
-                else {
-                    utilities_1.prettyLog(`Role assignment could not be created related to assignment id ${roleRequest.policyAssignmentId}. Status : ${response.httpStatusCode}`);
-                    message = response.content.error ? response.content.error.message : `Role Assignment could not be created. Status : ${response.httpStatusCode}`;
-                    status = exports.POLICY_RESULT_FAILED;
-                }
-                roleAssignmentResults.push(getRoleAssignmentResult(roleRequest.path, roleRequest.policyAssignmentId, roleRequest.policyDefinitionId, status, message));
-            });
-        }
-        catch (error) {
-            utilities_1.prettyLog(`An error occurred while creating role assignments. Error: ${error}`);
-            throw new Error(`An error occurred while creating role assignments. Error: ${error}`);
-        }
-    });
-}
-function getRoleDefinitionIds(policyDefinition) {
-    if (policyDefinition.properties
-        && policyDefinition.properties.policyRule
-        && policyDefinition.properties.policyRule.then
-        && policyDefinition.properties.policyRule.then.details
-        && policyDefinition.properties.policyRule.then.details.roleDefinitionIds) {
-        return policyDefinition.properties.policyRule.then.details.roleDefinitionIds;
-    }
-    return undefined;
 }
 function getPolicyDefinition(definitionPath) {
     const policyPath = path.join(definitionPath, exports.POLICY_FILE_NAME);
@@ -356,6 +226,7 @@ exports.getPolicyResults = getPolicyResults;
 function isCreateOperation(policyRequest) {
     return policyRequest.operation == exports.POLICY_OPERATION_CREATE;
 }
+exports.isCreateOperation = isCreateOperation;
 function validatePolicy(policy, path, type) {
     if (!policy) {
         throw Error(`Path : ${path}. JSON file is invalid.`);
